getting everything put together
