+++ pages/profile.js
@@ -1,5 +1,15 @@
+/* eslint-disable @next/next/no-img-element */
 import React from 'react';
+import { useAuth } from '../utils/context/authContext';
 
 export default function Profile() {
-  return <div>profile here</div>;
+  const { user } = useAuth();
+
+  return (
+    <>
+      <h1> {user.displayName} </h1>
+      <h1> {user.email} </h1>
+      <img alt="User Profile" src={user.photoURL} />
+    </>
+  );
 }
